+++ compiler-plugin/src/main/java/org.ballerinalang.java.jms/ServiceCompilerPlugin.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.ei.b7a.jms;
+package org.ballerinalang.java.jms;
 
 import org.ballerinalang.compiler.plugins.SupportedResourceParamTypes;
 import org.ballerinalang.compiler.plugins.AbstractCompilerPlugin;
