+++ utils/src/main/java/org/ballerinalang/java/jms/JmsDestinationUtils.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.ei.b7a.jms;
+package org.ballerinalang.java.jms;
 
 import org.ballerinalang.jvm.types.BTupleType;
 import org.ballerinalang.jvm.types.BType;
@@ -31,8 +31,6 @@ import javax.jms.Queue;
 import javax.jms.TemporaryQueue;
 import javax.jms.TemporaryTopic;
 import javax.jms.Topic;
-import java.util.ArrayList;
-import java.util.Objects;
 
 /**
  * Representation of {@link javax.jms.Destination} with utility methods to invoke as inter-op functions.
